Provide helper so app can be invoked directly from AWS Lambda
Use this method to service AWS Lambda requests from a custom Alexa skill directly without the need of an API Gateway. This hepler method will invoke the Flask Ask application by providing a WSGI-compatible environment that wraps the original Alexa event provided to the AWS Lambda handler. It returns the output generated by the Flask Ask application, which should be used as the return value.

Below is a full Flask-Ask application that can be run directly from an AWS Lambda function. Assuming your Lambda function is called 'lambda_function.lambda_handler', save this application to a file called 'lambda_function.py', package it with all of the Flask-Ask dependencies, and that's it.

    from flask import Flask
    from flask_ask import Ask, statement

    app = Flask(__name__)
    ask = Ask(app, '/')

    # This function name is what you defined when you create an
    # AWS Lambda function. By default, AWS calls this function
    # lambda_handler.
    def lambda_handler(event, _context):
        return ask.run_aws_lambda(event)

    @ask.intent('HelloIntent')
    def hello(firstname):
        speech_text = "Hello %s" % firstname
        return statement(speech_text).simple_card('Hello', speech_text)